requirements.txt
fastapi[standard]
asyncpg
sqlalchemy[asyncio]
python-dotenv
pyJWT
requests
google-auth
pandas